refactor(types): move types to their own file
